% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctAuto.R
\name{correctAuto}
\alias{correctAuto}
\title{Plate correction using autoregressive model}
\usage{
correctAuto(mat, diagWeight = 1/sqrt(2))
}
\arguments{
\item{mat}{Matrix presenting the data gathered from a plate. See details.}

\item{diagWeight}{The weight used for the colonize located diagonally of the focal colony. Default is one over the square-root of 2.}
}
\value{
A matrix with the same dimensions as \code{mat} but with the data corrected for edge effect problems.
}
\description{
Performs correction on a plate using an autoregressive model as a basis. In short, a autoregressive model is used where the eight closest neighbours of each colonie are used to make the correction (a queen pattern). It is highly recommended to standardize the data before performing the analysis. The function \code{\link{standardize}} was designed to perform plate standardization.
}
\details{
The argument \code{mat} is a matrix that has the same organisation as the plate to be corrected. 

In this function it is assume that closest neighbour of each colonie (up-down and left-right) are have the strongest influence and as such have a weight of 1, in this respect, the values given \code{diagWeight} have to be smaller or equal to 1. Conversely, a negative weight does not make sense either.
}
\author{
F. Guillaume Blanchet
}
\keyword{manip}
