% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctRC.R
\name{correctRC}
\alias{correctRC}
\title{Plate correction using a row-column standardization}
\usage{
correctRC(mat, lower = 0.25, upper = 0.75)
}
\arguments{
\item{mat}{Matrix presenting the data gathered from a plate. See details.}

\item{lower}{Numeric value presenting the lower fractional range of the data within each row (or column). This values needs to be range between 0 and 1. Default is 0.25. See details.}

\item{upper}{Numeric value presenting the upper fractional range of the data within each row (or column). This values needs to be range between 0 and 1. Default is 0.75. See details.}
}
\value{
A matrix with the same dimensions as \code{mat} but with the data corrected for edge effect problems.
}
\description{
Performs correction on a plate using a row-column standardization. Specifically, row means are calculated after having removed extreme values and each row is divided by this mean. Following, columns are standardized the same way.
}
\details{
The argument \code{mat} is a matrix that has the same organisation as the plate to be corrected. 

The argument \code{lower} needs to be always lower than the argument \code{upper} otherwise an error message will be sent.
}
\author{
F. Guillaume Blanchet
}
\keyword{manip}
