% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-data.R
\docType{data}
\name{plate}
\alias{plate}
\title{Plate data}
\format{
The data is organized as a data.frame that includes 2 columns, the first column are labels defining the location of each colonie on a plate while the second column present the raw value recorded during the experiment.
}
\usage{
data(plate)
}
\description{
Example of a 32 by 48 colonies plate that shows typical border boundary problems.
}
\details{
This data was provided by Jean-Philippe Côté (Université de Sherbrooke) and is organized following the typical structure the data is usually generated.
}
\keyword{datasets}
