% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Plate standardization}
\usage{
standardize(mat, FUN)
}
\arguments{
\item{mat}{Matrix presenting the data gathered from a plate. See details.}

\item{FUN}{Function to be used to carry out the standardization.}
}
\value{
A matrix with the same dimensions as \code{mat} but with the data standardized.
}
\description{
Standardize the data gathered for on a plate. This function is designed to apply any standardization on plate data.
}
\details{
The argument \code{mat} is a matrix that has the same organisation as the plate to be corrected. 

The function assumes that the standardization is performed on the whole data at once without reference to the matrix dimensions. In other words, the data is vectorized before performing the standardization.
}
\author{
F. Guillaume Blanchet
}
\keyword{manip}
